refactor(runtime-bootstrap): Move startup CLI and logging glue
Move NodeCli and LoggingConfigHandler into network.crypta.runtime.bootstrap with their package-local tests.

Update NodeStarter imports, expand the runtime.bootstrap package docs, and point SpotBugs exclusions at the relocated classes.
