+++ src/main/java/network/crypta/node/NodeStarter.java
@@ -28,6 +28,8 @@ import network.crypta.fs.AppDirs;
 import network.crypta.fs.AppEnv;
 import network.crypta.fs.Resolved;
 import network.crypta.fs.ServiceDirs;
+import network.crypta.runtime.bootstrap.LoggingConfigHandler;
+import network.crypta.runtime.bootstrap.NodeCli;
 import network.crypta.runtime.core.SSL;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.Logging;
